+++ CHANGELOG.md
@@ -28,7 +28,7 @@
 * Fix: [Android] add missing Android Foreground Service `phoneCall`.
 * Fix: [Android] reactive to microphone permission changes while call is ongoing.
 * Fix: [Android] failure while connecting to incoming/outgoing call with another call in-progress, `TVConnectionService` now checks for ongoing calls prior to stopping service. 
-* Feat: [Android] support null `From`/`To` parameters in raw `connect()` function.  
+* Feat: [Android, macOS] support null `From`/`To` parameters in raw `connect()` function.  
 * Feat: [Android] new FCM Token ignored on new token registration, now nowifies `twilio_voice` plugin of updated token.
 * Fix: [Web] code cleanup & refactor 
 * Fix: [iOS] remove unnecessary device token log when failed to register with Twilio.